[svn] * Fixed xp calculation for low level characters when grouped with a higher level player. Patch provided by Reiner030. * Fixed positive spells being not resistable when cast on hostile units. Patch provided by QAston. * Fixed compile warnings in gcc. Patch provided by WarHead.
--HG--
branch : trunk

+++ src/game/ArenaTeam.cpp
@@ -720,7 +720,7 @@ void ArenaTeam::UpdateArenaPointsHelper()
     if(stats.games < 10)
         return;
     // to get points, a player has to participate in at least 30% of the matches
-    uint32 min_plays = ceil(stats.games * 0.3);
+    uint32 min_plays = (uint32)ceil(stats.games * 0.3);
     for(MemberList::iterator itr = members.begin(); itr !=  members.end(); ++itr)
     {
         // the player participated in enough games, update his points
